1.5.2

+++ src/ring/fixed_msg_queue.c
@@ -24,7 +24,7 @@ fixed_msg_queue fixed_msg_queue_create(__in uint32_t one_msg_byte_size,
     fixed_msg_queue msg_queue_p = (fixed_msg_queue)raw_mem;
     msg_queue_p->one_msg_byte_size = one_msg_byte_size;
     msg_queue_p->ring_buf_p = RingBuffer_create_with_mem(raw_mem + sizeof(struct __fixed_msg_queue),
-        expect_mem_size - sizeof(struct __fixed_msg_queue));
+        (uint32_t)(expect_mem_size - sizeof(struct __fixed_msg_queue)));
     if (!msg_queue_p->ring_buf_p)
     {
         free(msg_queue_p);
